Add "list_column_types" method
Strip out all of the schema downloading stuff from casstcl.tcl.  We now entirely use the schema metadata traversing functions of the cpp-driger.  As this is kept up to date on the fly by the driver, changes in the schema will also be reflected in subsequent calls to the "list_schema", "list_tables", "list_columns" and "list_column_types" methods.

Uses a callout from C to Tcl to translate validator function names to data types, kind of cool.