New casstcl_obj_to_compound_cass_value_types
This function decodes a Cassandra data type expressed as a string in a Tcl object.  The special cases of list and set are two-element lists where the second element is the data type of the collection and the case of maps it's a three-element list where the second element is the data type of the keys in the map and the third element is the data type of the values.