new function casstcl_type_index_name_to_cass_value_types
This takes the type index name like wx.wx_metar.visibility and generates the 1, 2 or 3 CassValueType values depending on what kind of type it is.

What's extra cool is it maintains a hash table with the type index name as the argument and a casstcl_validatorHashInfo structure that it can return without variable lookups and parsing of any of it.