Create new Tcl obj type for caching cass types
Get rid of the Tcl hashtable that we were using to cache type conversions from like "int" to CASS_VALUE_TYPE_INT and "map text varint" and all that and do the caching using a new Tcl object type that we registered.

What's nice about this is that if the type map object is modified or deleted, the cached values will be recreated, versus with the hash table we were going to have to create a shadow way to invalidate the hash table yadda yadda, this way is much better.