cass-value-type-to-tcl-type bug fixes
You can't start overwriting fields in the Tcl object before you're done with it in its origin form, so in SetCassTypeTypeFromAny we now use local memory to store the values written by casstcl_obj_to_compound_cass_value_types and structure copy them into the tcl object if it's successful

Secondly we now provide a UpdateCassTypeString function to convert a cass value type from the internal rep to a string.  This is probably not needed and was cooked up from a misunderstanding of the cause of the above bug.