Replace three CassValueType with a structure
A number of routines took pointers to a primary and two secondary CassValueTypes.  Instead we now use a pointer to a casstcl_cassTypeInfo sturcture (which we already had) that comprises all three.

This simplified calling arguments to casstcl_obj_to_compound_cass_value_types, casstcl_bind_tcl_obj and casstcl_typename_obj_to_cass_value_types, and is overall a nice cleanup and simplification, especially in light of planned improvements coming to casstcl_make_upsert_statement, which was of the source of the insight that led to this commit.