Vastly improved error message
We were previously using cass_error_desc() to get the error description from a CassError value, but when we have a future result then a much more detailed error message is available using cass_future_error_message().

We now obtain that message and include it in the error message and errorCode when it is possible to.