make shared library end in .so on FreeBSD
By default it makes libcasstcl.so.1 but it installs it into /usr/local/lib/casstcl2.7 or whatever, anyway.  But for speedtables to find it it needs to be there as libcasstcl.so.  Sure we can symlink libcasstcl.so to libcasstcl.so.1 but the versioning doesn't matter anyway since the .so is installed with the specific version of casstcl it was built for.